fix(access): address CLI UX review feedback (swamp-club#678)
- access check exits 1 on DENY (implicit or explicit) for script/CI use
- grant revoke supports ID prefix matching (list shows truncated IDs)
- grant list --on help text clarifies exact-match behavior
- Extract buildModelMethodRunDeps to shared access_helpers.ts
- parseResourceFlag returns typed ResourceKind (eliminates cast)
- Renderers use writeOutput instead of console.log
- Consistent ellipsis on truncated grant IDs across renderers
