+++ src/cli/commands/access_grant.ts
@@ -31,30 +31,15 @@ import {
 } from "../repo_context.ts";
 import { UserError } from "../../domain/errors.ts";
 import { findDefinitionByIdOrName } from "../../domain/models/model_lookup.ts";
-import { resolveModelType } from "../../domain/extensions/extension_auto_resolver.ts";
-import { getAutoResolver } from "../auto_resolver_context.ts";
-import { DefaultMethodExecutionService } from "../../domain/models/method_execution_service.ts";
 import { modelRegistry } from "../../domain/models/model.ts";
 import { vaultTypeRegistry } from "../../domain/vaults/vault_type_registry.ts";
 import { reportRegistry } from "../../domain/reports/report_registry.ts";
-import { VaultService } from "../../domain/vaults/vault_service.ts";
-import { ExpressionEvaluationService } from "../../domain/expressions/expression_evaluation_service.ts";
-import { runFileSink } from "../../infrastructure/logging/logger.ts";
 import { GIT_SHA } from "./version.ts";
-import { join } from "@std/path";
-import {
-  SWAMP_SUBDIRS,
-  swampPath,
-} from "../../infrastructure/persistence/paths.ts";
-import { YamlDefinitionRepository } from "../../infrastructure/persistence/yaml_definition_repository.ts";
-import { SecretRedactor } from "../../domain/secrets/mod.ts";
-import { DataQueryService } from "../../domain/data/data_query_service.ts";
 import type { RepositoryContext } from "../../infrastructure/persistence/repository_factory.ts";
 import {
   consumeStream,
   createLibSwampContext,
   modelMethodRun,
-  type ModelMethodRunDeps,
 } from "../../libswamp/mod.ts";
 import { createModelMethodRunRenderer } from "../../presentation/renderers/model_method_run.ts";
 import {
@@ -64,133 +49,15 @@ import {
 } from "../../domain/models/access/grant_model.ts";
 import type { DataRecord } from "../../domain/data/data_record.ts";
 import { createAccessGrantListRenderer } from "../../presentation/renderers/access_grant.ts";
+import {
+  buildModelMethodRunDeps,
+  LOCAL_PRINCIPAL,
+  parseActionsFlag,
+  parseResourceFlag,
+} from "./access_helpers.ts";
 
 // deno-lint-ignore no-explicit-any
 type AnyOptions = any;
-
-const LOCAL_PRINCIPAL = "user:local";
-
-export function parseResourceFlag(
-  value: string,
-): { kind: string; pattern: string } {
-  const colonIdx = value.indexOf(":");
-  if (colonIdx === -1) {
-    throw new UserError(
-      `Invalid resource selector "${value}": expected format "kind:pattern" (e.g. "workflow:@acme/*")`,
-    );
-  }
-  const kind = value.substring(0, colonIdx);
-  const pattern = value.substring(colonIdx + 1);
-  const validKinds = ["workflow", "model", "data", "access"];
-  if (!validKinds.includes(kind)) {
-    throw new UserError(
-      `Invalid resource kind "${kind}": must be one of ${
-        validKinds.join(", ")
-      }`,
-    );
-  }
-  return { kind, pattern };
-}
-
-function parseActionsFlag(value: string): string[] {
-  const actions = value.split(",").map((a) => a.trim()).filter((a) =>
-    a.length > 0
-  );
-  const validActions = ["run", "read", "write", "admin"];
-  for (const action of actions) {
-    if (!validActions.includes(action)) {
-      throw new UserError(
-        `Invalid action "${action}": must be one of ${validActions.join(", ")}`,
-      );
-    }
-  }
-  if (actions.length === 0) {
-    throw new UserError("At least one action is required");
-  }
-  return actions;
-}
-
-function buildModelMethodRunDeps(
-  repoDir: string,
-  repoContext: RepositoryContext,
-  isDirectExecution: boolean,
-): ModelMethodRunDeps {
-  return {
-    repoDir,
-    lookupDefinition: (idOrName) =>
-      findDefinitionByIdOrName(repoContext.definitionRepo, idOrName),
-    getModelDef: (type) => resolveModelType(type, getAutoResolver()),
-    createEvaluationService: () => {
-      const dqs = new DataQueryService(
-        repoContext.catalogStore,
-        repoContext.unifiedDataRepo,
-      );
-      return new ExpressionEvaluationService(
-        repoContext.definitionRepo,
-        repoDir,
-        {
-          dataRepo: repoContext.unifiedDataRepo,
-          dataQueryService: dqs,
-        },
-      );
-    },
-    loadEvaluatedDefinition: (type, name) =>
-      repoContext.evaluatedDefinitionRepo.findByName(type, name),
-    saveEvaluatedDefinition: (type, definition) =>
-      repoContext.evaluatedDefinitionRepo.save(type, definition),
-    createExecutionService: () => new DefaultMethodExecutionService(),
-    createVaultService: () => VaultService.fromRepository(repoDir),
-    dataRepo: repoContext.unifiedDataRepo,
-    definitionRepo: repoContext.definitionRepo,
-    outputRepo: repoContext.outputRepo,
-    dataQueryService: new DataQueryService(
-      repoContext.catalogStore,
-      repoContext.unifiedDataRepo,
-    ),
-    createRunLog: async (modelType, method, definitionId) => {
-      const redactor = new SecretRedactor();
-      const timestamp = new Date().toISOString().replace(/[:.]/g, "-");
-      const logFilePath = join(
-        swampPath(repoDir, SWAMP_SUBDIRS.outputs),
-        modelType.normalized,
-        method,
-        `${definitionId}-${timestamp}.log`,
-      );
-      const logCategory: string[] = [];
-      await runFileSink.register(
-        logCategory,
-        logFilePath,
-        redactor,
-        swampPath(repoDir),
-      );
-      return {
-        logFilePath,
-        redactor,
-        cleanup: () => runFileSink.unregister(logCategory),
-      };
-    },
-    createAndSaveDefinition: isDirectExecution
-      ? async (type, definition) => {
-        const autoDefRepo = new YamlDefinitionRepository(
-          repoDir,
-          undefined,
-          swampPath(repoDir, SWAMP_SUBDIRS.autoDefinitions),
-          false,
-        );
-        await autoDefRepo.save(type, definition);
-      }
-      : undefined,
-    getDefinitionPath: isDirectExecution
-      ? (type, id) => {
-        return join(
-          swampPath(repoDir, SWAMP_SUBDIRS.autoDefinitions),
-          type.toDirectoryPath(),
-          `${id}.yaml`,
-        );
-      }
-      : undefined,
-  };
-}
 
 export async function queryGrants(
   repoContext: RepositoryContext,
@@ -370,7 +237,7 @@ const accessGrantListCommand = new Command()
     "Repository directory (env: SWAMP_REPO_DIR)",
   )
   .option("--subject <subject:string>", "Filter by subject")
-  .option("--on <resource:string>", "Filter by resource selector")
+  .option("--on <resource:string>", "Filter by resource selector (exact match)")
   .action(async function (options: AnyOptions) {
     const ctx = createContext(options as GlobalOptions, [
       "access",
@@ -437,10 +304,16 @@ const accessGrantRevokeCommand = new Command()
     ]);
 
     const allGrants = await queryGrants(repoContext);
-    const match = allGrants.find((r) => r.grant.id === grantId);
-    if (!match) {
+    const matches = allGrants.filter((r) => r.grant.id.startsWith(grantId));
+    if (matches.length === 0) {
       throw new UserError(`Grant not found: ${grantId}`);
     }
+    if (matches.length > 1) {
+      throw new UserError(
+        `Ambiguous grant ID prefix "${grantId}" — matches ${matches.length} grants. Use a longer prefix.`,
+      );
+    }
+    const match = matches[0];
     if (match.grant.state === "revoked") {
       ctx.logger.info`Grant ${grantId} is already revoked`;
       return;
