+++ src/presentation/renderers/access_group.ts
@@ -19,6 +19,7 @@
 
 import type { Group } from "../../domain/models/access/group_model.ts";
 import type { OutputMode } from "../output/output.ts";
+import { writeOutput } from "../../infrastructure/logging/logger.ts";
 
 export interface AccessGroupListRenderer {
   renderList(groups: Group[]): void;
@@ -32,44 +33,44 @@ function formatPrincipal(p: { kind: string; id: string }): string {
 class LogAccessGroupListRenderer implements AccessGroupListRenderer {
   renderList(groups: Group[]): void {
     if (groups.length === 0) {
-      console.log("No groups found.");
+      writeOutput("No groups found.");
       return;
     }
 
     const header = `${"NAME".padEnd(30)}  ${"MEMBERS".padEnd(8)}  ${
       "CREATED BY".padEnd(20)
     }  CREATED AT`;
-    console.log(header);
+    writeOutput(header);
 
     for (const group of groups) {
       const name = group.name.padEnd(30);
       const members = String(group.members.length).padEnd(8);
       const createdBy = formatPrincipal(group.createdBy).padEnd(20);
       const createdAt = group.createdAt.slice(0, 10);
-      console.log(`${name}  ${members}  ${createdBy}  ${createdAt}`);
+      writeOutput(`${name}  ${members}  ${createdBy}  ${createdAt}`);
     }
   }
 
   renderMembers(group: Group): void {
     if (group.members.length === 0) {
-      console.log(`Group "${group.name}" has no members.`);
+      writeOutput(`Group "${group.name}" has no members.`);
       return;
     }
 
-    console.log(`Members of "${group.name}":`);
+    writeOutput(`Members of "${group.name}":`);
     for (const member of group.members) {
-      console.log(`  ${formatPrincipal(member)}`);
+      writeOutput(`  ${formatPrincipal(member)}`);
     }
   }
 }
 
 class JsonAccessGroupListRenderer implements AccessGroupListRenderer {
   renderList(groups: Group[]): void {
-    console.log(JSON.stringify(groups, null, 2));
+    writeOutput(JSON.stringify(groups, null, 2));
   }
 
   renderMembers(group: Group): void {
-    console.log(JSON.stringify(group, null, 2));
+    writeOutput(JSON.stringify(group, null, 2));
   }
 }
